Move datastore impl from owned object index module to store
